ipn/ipnlocal: handle traffic-steering when feature is not enabled
In PR tailscale/corp#34401, the `traffic-steering` feature flag does
not automatically enable traffic steering for all nodes. Instead, an
admin must add the `traffic-steering` node attribute to each client
node that they want opted-in.

This PR handles the situation where an admin has assigned this node
attribute to a client, but the tailnet does not have this feature
enabled. When this happens, the client will fall back to the previous
DERP-based algorithm.

Updates tailscale/corp#34399

Signed-off-by: Simon Law <sfllaw@tailscale.com>