tsnet: respond quickly to state changes in auth loop
tsnet has a 5s sleep as part of its logic waiting for login to succeed.
Add an additional channel that will interrupt this sleep early if the
local backend's state changes before then. This is early enough in the
bootstrap logic that the local client has not been set up yet, so we
subscribe directly on the local backend in keeping with the rest of the
function, but it would be nice to port the whole function to the new
eventbus in a separate change.

This was noticed while debugging what makes tsconsensus tests slow, but
it probably affects other test suites as well.

Updates #16340

Change-Id: I7a28fd3927bbcdead9a5aad39f4a3596b5f659b0
Signed-off-by: Tom Proctor <tomhjp@users.noreply.github.com>