wgengine: delay disco key exchange when control is unavailable
Instead of addig a delay to the exchange of disco keys to avoid a race
with the wireguard handshake, look for the handshake itself and ensure
that we do not start the exchange when the handshake is ongoing.

We do not have a direct way of knowing that a handshake is ongoing, only
how many are currently in flight or have failed. That makes the polling
less straight forward.

Ideally, we would have wg tell us when a handshakes is ongoing so we can
avoid polling for it.

Updates tailscale/corp#34037

Signed-off-by: Claus Lensbøl <claus@tailscale.com>