.github: speed up cigocacher build with a small bootstrap cache
This reduces the time to build cigocacher from 25-30s to ~5s including
the cache restore step, as the cache is only ~20MB.

I also discovered while adding this step that the **/go.sum pattern was
taking 5-15s to evaluate because it's traversing the whole checkout tree
structure, but we only have one go.sum file in the repo that we want to
do content cache busting for. Replacing ** with * means we only look one
level deep for go.sum (in the src/ checkout dir), and knocks 5-15s off
every job.

Updates #10808

Change-Id: I6ccb0c53593e0ccb4048734bb3a0624e23bd17f2
Signed-off-by: Tom Proctor <tomhjp@users.noreply.github.com>