cmd/k8s-operator: add reconciler for Tailnet resource
This commit adds the new reconciler for the Tailnet resource in a new
location. This reconciler follows a possible new package structure
within `/k8s-operator/reconciler/tailnet`. It also varies from our
existing reconciler in using a single constructor function that creates
the reconciler and registers it with the controller manager.

This is a potential new structure for us to follow for adding new
reconcilers and a style we can migrate our existing reconcilers into.

Fixes: https://github.com/tailscale/corp/issues/34767
Signed-off-by: David Bond <davidsbond93@gmail.com>