net/dns: add CnameHosts field to Config
A new feature in DNS Config supports CNAME records. It allows users to
link FQDNs, even those with wildcards, to different destination FQDNs.
Although the FQDNs in question might be associated with CNAME names, a
final resolution to an A or AAAA address is required by the DNS server

Updates #1196

Signed-off-by: Fernando Serboncini <fserb@tailscale.com>