net/dns/resolver: return all matching A or AAAA records
The wire protocol already supports receiving multiple DNS records for a
name from the server, but so far we only return one of those records in
our DNS resolver. This patch updates the resolver to return all matching
A or AAAA records (or both, for an ALL request), depending on the
requested type.

Headscale supports configuring extra DNS records with more than one
record for a given name. This patch has been tested end-to-end with such
a configuration.

Updates #4324
Updates #5313

Signed-off-by: Luc Ritchie <luc.ritchie@gmail.com>