tool: add cigocacher
Adds tool script for downloading and running cigocacher from releases
based on a pinned revision. This is only designed to run in CI, so we
can safely depend on git bash being installed like it is by default on
GitHub's hosted Windows runners, and that saves us from having to
maintain multiple translated copies of the same script.

Updates tailscale/corp#10808

Change-Id: I9a84b6b83ff145f0860b5d21c59cb5c15e9bed0b
Signed-off-by: Tom Proctor <tomhjp@users.noreply.github.com>