net/dns/resolver: Truncate UDP DNS responses
Perform RFC-compliant truncation of DNS responses
which are to be delivered over UDP when they
exceed the maximum allowed size.

The maximum size is determined by the EDNS0
options in the DNS request, or defaults to
512 bytes if no EDNS0 options are present (or
the ENDS0 value in the request is less than 512).
Per DNS Flag Day 2010 recommendations, we cap
the maximum UDP DNS response size at 1232 bytes.

Fixes #13601 and #18107

Signed-off-by: Ryan Koski <ryan@koski.org>