k8s-operator: add "tailnet" field to custom resources (#18174)
This commit modifies our `Connector`, `ProxyGroup` and `Recorder` resources
by adding a new field named "tailnet" that will be used to specify the
set of oauth credentials that should be used to generate the auth key
that authenticates them.

This field is optional and when set should contain the name of an existing
`Tailnet` resource.

Updates: https://github.com/tailscale/corp/issues/35132

Signed-off-by: David Bond <davidsbond93@gmail.com>